pyproject.toml
[project]
name = "agentic-clearinghouse"
version = "0.1.0"
description = "The Agentic Clearinghouse - Escrow & verification protocol for AI agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Redis
    "redis[hiredis]>=5.2.0",
    # State Machine (transition guard)
    "python-statemachine>=2.5.0",
    # Orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    # Verification
    "e2b-code-interpreter>=1.0.0",
    "litellm>=1.50.0",
    # Payments
    "coinbase-agentkit>=0.7.0",
    # MCP Protocol
    "mcp[cli]>=1.0.0",
    # Observability
    "structlog>=24.4.0",
    # Utilities
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "aiosqlite>=0.22.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "factory-boy>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_clearinghouse"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["B008"]  # Depends() in FastAPI argument defaults is idiomatic

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests that hit real external APIs (E2B, Gemini)",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
